requirements.txt
langchain
openai
pypdf
requests
beautifulsoup4
python-dotenv
fastapi
uvicorn
